move limit to `rustc_ast_ir`
